removed analysis page
