+++ src/components/Navigation/Navigation.js
@@ -1,0 +1,9 @@
+import NavigationLinks from './NavigationLinks';
+
+export default function Navigation() {
+	return (
+		<nav>
+			<NavigationLinks />
+		</nav>
+	);
+}
